compatibility: add use case for library server
We want to be able to check software compatbility, amongst other things.
In Kubernetes, Node Feature Discovery (NFD) has a design where a daemon
runs on the nodes, can parse what they provide, and then provides
that to a central service. For our case, we can do similar - having
a service (that can run on the node and either work with a local client
OR a scheduler) that knows how to read either a compatibility artifact
directly (json payload) or retrieve from a registry, where it describes
an application or container, and then based on the libraries needed,
quickly determine if the node can satisfy the needs. This I am calling
a Compatibility server and check because it extends to other things
beyond libraries / software.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>